# openapi.yaml
openapi: 3.1.0
info:
  title: Flyleaf API
  description: The social reading tracker backend (PRD §24, Architecture §6).
  version: 0.0.1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT (FN-63)
...
paths:
...
  /admin/dedupe/queue:
    get:
      summary: List dedupe review queue
      tags:
        - Admin Dedupe
      description: Lists pending or resolved Stage 3 & 4 duplicate candidates.
      parameters:
        - schema:
            type: string
            enum:
              - pending
              - merged
              - dismissed
            default: pending
          in: query
          name: status
          required: false
        - schema:
            type: integer
            enum:
              - 3
              - 4
          in: query
          name: stage
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          in: query
          name: limit
          required: false
        - schema:
            type: integer
            minimum: 0
            default: 0
          in: query
          name: offset
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        stage:
                          type: integer
                          enum:
                            - 3
                            - 4
                        status:
                          type: string
                          enum:
                            - pending
                            - merged
                            - dismissed
                        confidence:
                          type:
                            - "null"
                            - number
                        reason:
                          type: string
                        dismiss_reason:
                          type:
                            - "null"
                            - string
                        created_at:
                          type: string
                          format: date-time
                        reviewed_at:
                          type:
                            - "null"
                            - string
                          format: date-time
                        reviewed_by_user_id:
                          type:
                            - "null"
                            - string
                          format: uuid
                        survivor:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            title:
                              type: string
                            authors:
                              type: array
                              items:
                                type: string
                            first_publish_year:
                              type:
                                - "null"
                                - integer
                            log_count:
                              type: integer
                            edition_count:
                              type: integer
                            reads_count:
                              type: integer
                            cover_id:
                              type:
                                - "null"
                                - integer
                          required:
                            - id
                            - title
                            - authors
                            - log_count
                            - edition_count
                        loser:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            title:
                              type: string
                            authors:
                              type: array
                              items:
                                type: string
                            first_publish_year:
                              type:
                                - "null"
                                - integer
                            log_count:
                              type: integer
                            edition_count:
                              type: integer
                            reads_count:
                              type: integer
                            cover_id:
                              type:
                                - "null"
                                - integer
                          required:
                            - id
                            - title
                            - authors
                            - log_count
                            - edition_count
                      required:
                        - id
                        - stage
                        - status
                        - reason
                        - created_at
                        - survivor
                        - loser
                required:
                  - data
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
...
  /admin/dedupe/queue/{id}/resolve:
    post:
      summary: Resolve dedupe candidate
      tags:
        - Admin Dedupe
      description: Approves merge or dismisses duplicate candidate with reason.
        Requires administrator role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum:
                    - merge
                    - dismiss
                reason:
                  type: string
              required:
                - action
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  action:
                    type: string
                    enum:
                      - merge
                      - dismiss
                  merge_id:
                    type: string
                    format: uuid
                required:
                  - success
                  - action
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        description: Machine-readable error code
                      message:
                        type: string
                        description: Human-readable error description
                      field:
                        type: string
                        description: Invalid field name if applicable
                    required:
                      - code
                      - message
                required:
                  - error
...
servers:
  - url: /v1
    description: Default API path
...